// src/plugins/character-creator/shared/interfaces.ts
export interface IInternal {
    primaryNav: number;
    secondaryNav: number;
    modalOpen: boolean;
    motherIndex: number;
    fatherIndex: number;
    hairIndex: number;
    shoesIndex: number;
    legsIndex: number;
    topsIndex: number;
    hatIndex: number;
    watchIndex: number;
    braceletIndex: number;
    earingIndex: number;
    glassesIndex: number;
    undershirtIndex: number;
}